qa_test/requirements.txt
# 核心依赖
numpy>=1.21.0
pyyaml>=5.4.1
tqdm>=4.62.0

# 大语言模型API
openai>=0.27.0
anthropic>=0.3.0
requests>=2.28.0

# NLP处理
jieba>=0.42.1
nltk>=3.7
rouge-score>=0.1.2
textstat>=0.7.3

# 语义分析
sentence-transformers>=2.2.0
torch>=1.10.0
scikit-learn>=1.0.0

# 网络重试
tenacity>=8.0.0

# 可选依赖（用于高级功能）
# transformers>=4.20.0  # 如果需要使用更多预训练模型
# pandas>=1.3.0  # 如果需要数据分析功能
# matplotlib>=3.4.0  # 如果需要可视化功能